New JSON key logic
This commit adds a few nice json key processing
- Finally a working itemlist computed property that Solr can see. Had
to work around the computing, since Solr uses iterators instead of
getValue()
- A JSon Path generator capable of detecting lists and named properties
to add wildcards to paths. Used as a computed property in any
strawberry field
- A new Vocabulary deployed on install
- On Node preserve, Keys are extracted as JSON Paths, split back and
Taxonomy terms, with cool! Hierarchies are created auto-magically.

Implementation works, indexes in Solr but better than that, potential
of this is 345%