Pretty print JSON when editing RAW
This is just a tiny help for people editing raw JSON. Will try to decode first. If that works, it will encode back with Pretty print enabled. If it does not work, which is only a case for a corrupt DB, because we have Symfonfy Constraints in place when saving and also because our DB Schema is set for JSON (means we will never get there) output the original, not pretty print so people could eventually fix it.