Make sure we have a default widget during field instance setup
# FIXES

Main issue here is how Drupal 8 decides which widget to show when setting a field inside the Content Type Admin menu. There you can set default values, etc, but if a special $form_state key is not present *which can be only set via a form alter and is not used anywhere i core* if will render the widget used in the default form display to set The defaults. That is in our case totally wrong, we don't want to set a default JSON via a webform in this context (and it also does not work).

# IMPROVES
Since our StrawberryField Field uses mostly computed Properties i wanted to improve computing times. We now use much more ::computed to check instead of going into ::processed content on each property. There are tiny changes here, but it will drop a few extra cycles if the main property of the field is empty (->value) and will also be way more responsive on calculating if the field is empty which seems something Drupal does a lot.