Make a StrawberryField Item Normalizer
This pull allows our JSON encoded Strawberryfield to be expanded into a full array (and then serialized as it needs to) for REST, API and in general Symfony and Drupal serialization needs. This is another step into having a fully functional strawberry seeds module. Now we can export Full Objects with all its JSON