Adds a plain JSON  processing event
The main ideas to to be able to Dispatch a Special Event type that contains a JSON Array with raw values (full content of a SBF or a webform input), a second JSON Array with data that needs to be processed (could be a snippet, piece of it, etc) and is able to  store back into the event the changes that happen to the last one while Subscribers play with it.

I'm having my doubts right now if passing an arrays here instead of JSON encoded Objects are the right way, Arrays are easier to manage to people making changes to the data, but Objects have the benefit that i don't need to write the changes back to Event since PHP keeps a pointer to the original OBJECT.

Sadly if i go JSON ENCODE OBJECT i will force people to decode, mangle and reencode and in that case the reference/pointer is of no use anymore.

Another option would be to create a special Custom Object that has both arrays as elements. That way i can still access the arrays directly and write them back. Feels a bit convolued so i'm going now for just Arrays. @giancarlobi ideas? What do you think?