Fix on DataSource
I found out that our DataSource was throwing a error on Content/Node Save which was weird and quite bad.

Happents that \Drupal\strawberryfield\Plugin\search_api\datasource\StrawberryfieldFlavorDatasource::getItemId() is called on entity save (node) because the tracker (which is a procedural call on the search_api.module) uses that to set which documents need to be tracked. So at that stage there is no item_id yet present and the node has not that one.
This commit fixes that by providing a fake (for now) single valued item id based on the node id, the language and the first sequence (0). But we will still need to add to that all the other pages, if any exist, to the tracker.
See https://git.drupalcode.org/project/search_api/blob/8.x-1.x/search_api.module#L355-361

We should discuss what to do here. We can either check to which index we are going to be added, and invoke that manually for every page (if any exist) for that entity or we can simply return NULL too and deal with tracking and everything here ourselves.