Refactor to allow more KeyName Providers
@giancarlobi this is a big refactor for you to look at, still WIP, but should be mostly solved this weekend.

Note: Refactor does not affect people's data (hey! yes!) and is only made to allow more type of Key Name Providers+ Flavours and a better experience writing new ones as we advance.

The supplied Flavor key Name provider is mostly Dummy right now. It just exposes to the outside and validates the internal Services via JSON Schema Draft 7.0 validation and sees if the file can be loaded, but that is  more than enough. Idea here is to make the actual ZIP processing  its own helper class that is called inside the processor. That processor will just go and read the manifest. DataSource then will use that to index the content. SO we are pretty close.

This pull also has JSONAPI drop file integration *basic still, allows to upload, not to fetch yet.

There is also a big piece of cleaning, more verbosity and smaller corrections.

See all the `@TODOS`. Will be solving the one by one this weekend

Until this is merged into Master you need to

composer require swaggest/json-schema in your base D8 installation. Once this gets merged the updated local composer.json will kick in.