Deposit your Digital Objects to Storage via Event Subscriber!
You like having DO in file instead of just DB? Want to use static file generators with source data that is complete and easy to parse?
You feel safer at night if your DO are being backuped as files too my the MACHINA?
well this does that. And also generates a .gz version (smaller).
Fun stuff. a digital object is 6 Kbytes long... jajaja
There is space for improvement, and in the future we can plug in here different File storage layouts (like using our OCFL class finally). @giancarlobi to discuss what we want to dump. `@TODO`: deal with versions (a different subscriber or same but with some if/elses)
Side effeeeeeect. This triggers before save to DB. So if something goes wrong? Light goes out. You have a backup =)