Update Event Subscriber Services
Adds the StrawberryfieldEventInsertFileUsageUpdater service which kicks in to track/update usage for just new inserted nodes.
@giancarlobi something new/side effect. Since i added the new field_file_drop field we also have automatic file tracking for every node! Its just one time but enough to avoid temporaries to be removed

Updates StrawberryfieldEventInsertSubscriberDepositDO to deposit JSON representation of the full node + a slimmer, SBF one with some extra elements (we can discuss how useful those are or not). A similar one will deal with versions.

Removed return for