Fix edge case introduced by new 'field_file_drop' field
This is new: When you are saving a Strawberryfield field with wrong JSON (means you edited it manually) the 'field_file_drop' field compute kicks in before the SBF is validated (means it holds a wrong JSON temproary in its getValue() response). Normally the TypeData validation kicks in in time before save happens, but no before assignment happens and since 'field_file_drop' tries to computed the flattened structure (which of course can not happen if the JSON is wrong and we can not enforce the order of compute) we get an error. By forcing a local validation first before flattened properties are requested and this is computed we ensure that even in this not-yet-persisted state we can safely run the whole normal cycle. Now Typedata actually runs its validation and we get a Wrong JSON error and all works normal. @giancarlobi i will have to watch this closer in the future since even if i know this is the only place where this happens, we need to make sure that invalid data is never accesed, even when we are dealing with not yet saved to DB data, specially since we require a lot of precomputed properties.