Try to make frictionlessdata/datapackage a dependency.
I also asked in that github account of a release (right now i had to target "frictionlessdata/datapackage": "dev-master") since the current 0.x release uses alchemy/zippy 0.3.5 because of bug but drupal console requires 0.4.3 and there is conflict. I would really love to use this package instead of parsing manually the JSON, but if there is no future release i will be forced to simply use my own code to deal with accessing the ZIP file and parsing the datapackage.json file.