Better file name sanitization
This deals with UTF-8, ascii remapping, stripping and replacing almost everything with a '-'. I also checked AWS S3 limits and its 512 chars long for a filename, so cutting the final name to that. We also add thew mimetype as a prefix + the UUID as a suffix, to totally avoid any clashing of files. The suffix allows quick filter in S3 enabled platforms but also a good indicator for endusers navigating filesystem directly.