Add PHP doc and return local property if set
No need to recalculate if the value was set on construct. Since Service Containers are singletons, means a single instance during the life cycle of a request, we don't need to calculate over and over this function. To be honest, i feel we should not even initialize it, and only leave it as a on demand public method. But for the sake of doing a single thing at a time, i will leave this and revisit in beta3