WIP
Bring EXIFTOOL and FIDO into place
Tested and works. Good. Still requires this to pass via the webform since we have not enabled the service to kick in as an independent process after webform generation (we should... i have an issue open). But it does generate the needed data and flooods the JSON with normally not very useful info