WIP still but better.
I need to test this but there are some chanes
Since service containers are singletons (means initialized once per request only)
So we can check, if needed here if fido and exif are well formed. I guess i will end removing this and maybe this can go into a STATUS CHECK? in global Drupal, like when you need a security patch. Ok, for now this is a mix of safe and performant. If both are wrong AND toogle is enabled it will logg it also.
- Also, it checks now if the file is already in temporary storage before trying to move it. Imagine some consecutive operations because someone is just testing, no need to move 2TBs of video back and forth right?
- Also a message fix.
- I think there is still a bug when the uploaded file's basename is not a URL encoded...