Adds permissions to View RAW JSON metadata
Now that we extract EXIF and there can be a lot of fun and private stuff in EXIF data we can / should allow a permission on the RAW display

This adds 2 new permissions that only apply to the StrawberryDefaultFormatter
View any raw JSON stored in a Strawberryfield
View own raw JSON in a Strawberryfield

This is how things work here:
- if the entity (node, file, etc) has an owner the 'Own' can be actually evaluated
- every user can have the 'View any' of course
- Every user that can edit an entity can see the RAW metadata (i see no benefit on hiding data with editing capabillities to be honest)
Users with 'bypass node access' can see everything as always.

TODO: of course a lot, we can maybe make this a more general method. I can also inject the user here instead of using the drupal:: service. But then, i don't think its crucial here right now.