Tiny update on automatic ordering
 This code here now preserves manual order (whatever/however that was generated, which includes also untouched automatic ordering, we can not know how odering happened, just that it happened). But if there is ordering in place and we add new files, it will simply push new files (ordered by whatever order automatically is already happening) to after the current ones. I feel this is fair.

I will then, once this is proven to be right program a flag: basically a 'command' passed via  webform element (key and value) or via just json: that can affect this. Command can be 'force reordering' or .. wait of it, force EXIF, etc. I wanted to do this for long time. Now its the (for 1.0.0.) @giancarlobi once i merge this into ISSUE-86, can you test? Thanks!!