Remove escapeshellarg
I should never trust forums, i had this right!
TODO: we need to move this curl calls (simple yes, but hard to parse the response) to symfony http client calls. Why? We need to know the actual response of an upload so we can decide how to proceed. Example? 0 Bytes uploaded because the file path was incorrectly single quoted!