Invert multiple to single to allow VBO use case to be easier to apply
So here i'm doing some crazy code recycling. This action is basically of the type the admin configures onces and can be triggered somewhere. But in AMI (github.com/esmero/ami) we extend the class to be of VBO (Views Bulk Operations type) where the same methods used for once configuration become the User facing form (weird...who did this?) and now the configuration by the admin goes somewhere. Since VBO uses the single execution in loop for the multiple one, we changed the original class signature. And it seems to work!

Also fixes the module check (i need help folks for realz!), was never named codemirror, its named codemirror_editor