Makes Flavour Data Source actually work. Of course it needs runners
So someone can feed it. But It actually indexes! This is a WIP of course, but it has a few benefits right now as it is: no longer tracked automatically since the number of IDs we will get really depends dynamically on what the NODES have (files, data) so we can not have a fixed number. SB Runners provides the actual entries directlty via QUEUE worker (smooth)

There are a few TODOS here and some cleanup (remove logs) but as it is we can already index higlights for PDF extractions, HOCR, WARC urls, etc and those can be displayed not as rendered entities but as fields (highlight + title linking back to the original Entity). iF we add some extra logic we can pass also URL fragments for the book reader!