Issue 101: JSON Patching, D9 fixes and new JSON Key name provider (#105)
* baby steps into the JSON PATCH action

Still in debug mode here, so not modifying yet the content.
Some issues i see are that JSON Pointers have NO wildcards.
So complex selections like the ones we do normally in JMESPATH or JSONPATH are not possible. WHY? but WHY?
Ok, i was about to cry but then i remembered i know how to code!
SO next step is, yeah, put your JMESPATH compatible path instead of a JSONPointer there (i will just check for * and dots and []. If i find that, i will evaluate those against each JSON, will get the ACTUAL real paths that resolve to those * and create on the FLY (yes! like when you code) new RFC spec correct JSONPOINTERS (statics) ones. So cool, Yeah. so cool

* More debug/ First time this works, yay!

I was getting some mixed results but now its working fine. Will leave dpm() until i finish coding

* getAccountName() replaces getUserName

Deprecated property my D9 checker missed. @giancarlobi thanks!

* Better JSON Patch without a form yet

Next pass now is to allow the user to add custom JSONPATCH JSON and add extra info like Where to start to patch, and also hopefully a JMESPATH preprocessor but without casting (basically a subset that allows * instead of a given array key or Object ID in the path, given the fact that JSONPATCH only works with JSONPOINTERS

* Invert multiple to single to allow VBO use case to be easier to apply

So here i'm doing some crazy code recycling. This action is basically of the type the admin configures onces and can be triggered somewhere. But in AMI (github.com/esmero/ami) we extend the class to be of VBO (Views Bulk Operations type) where the same methods used for once configuration become the User facing form (weird...who did this?) and now the configuration by the admin goes somewhere. Since VBO uses the single execution in loop for the multiple one, we changed the original class signature. And it seems to work!

Also fixes the module check (i need help folks for realz!), was never named codemirror, its named codemirror_editor

* Adds identify and pdfinfo config options and executes limited identify

This adds two new binaries to the basic, non background, file preprocessing and a form element to set the paths up (shipped with our docker deployment).

I'm not so happy with the Identify processing speed, but also my computer is very slow.

All this needs to go and move into SB Runners once i have time and brain clarity to do so since large PDFs, Gigabyte long movies, etc, will load just too slow and its better if we do this all in batch background via our SBF Runners Queue.

* Global Entity Reference Field that covers any relationship between objects

@giancarlobi for your eyes

* And i got it working! SBF Entity Reference Properties!

@giancarlobi @alliomeria i think i got it :ocean: :palm_tree:  :sunny: :surfer: !!
This is of course a first approach but it is wonderful to see

I got Parent's Collection't title indexed into the Child Solr Document, also its RDF type and even the parent of a parent!

This stuff is good. But i need more testing and some Drupal 9 pass, plus better options. Right now its totally fixed on NODE entities (can be terms in the future too) and depends on adding a new Key name provider using ismemberof as the JMESPATH. (new Entity JMESPATH provider). Config entity Schema also provided!

* Add missing  config_export Annotation item for the  Key Name Provider Entity

@giancarlobi can you test this please. I hope i did not miss any of the keys

* Makes Flavour Data Source actually work. Of course it needs runners

So someone can feed it. But It actually indexes! This is a WIP of course, but it has a few benefits right now as it is: no longer tracked automatically since the number of IDs we will get really depends dynamically on what the NODES have (files, data) so we can not have a fixed number. SB Runners provides the actual entries directlty via QUEUE worker (smooth)

There are a few TODOS here and some cleanup (remove logs) but as it is we can already index higlights for PDF extractions, HOCR, WARC urls, etc and those can be displayed not as rendered entities but as fields (highlight + title linking back to the original Entity). iF we add some extra logic we can pass also URL fragments for the book reader!

* Not sure if we got this wrong before? But it looks OK now. So yeah..

@giancarlobi if you can test if mime types are being guess correctly that could help a lot

* Never used but those Event Subscribers looked off

This are abstract classes made for the future, but better have them using the right events. Felt a little bit like own copy pasta that neede a fix