Create OcrHighlight.php
This may be removed. We do not need it anymore since we have the config but i left it so i can discuss with @giancarlobi and show him how this interacts.

the main issue with this is that you can create anything (great!) but the backend class can only read a few. Like a little few!