Get removed file ids and call the cleaner during presave
The logic here is simple
We ask SBF for its original (previous saved data) and get the flattened values
We ask for all the dr:fid values (means any file mentioned/referenced)
During new file processing we keep a list of all fids added
After that we compare old files with new files. Difference is Magically (hahaha) the ones that are not longer with us.
We call  $this->strawberryfilepersister->removefromAsFileStructure(
                $to_be_removed_files, $fullvalues, $entity->id()); to remove those left over as: keys from the JSON
Then instead of iterating and cleaning over the "newly processed" because guess what.. if you deleted all newly processed will have no iteration, we iterate over every possible key (only a few StrawberryfieldJsonHelper::AS_FILE_TYPE really) and check if there is previous data for that as:key, if there is new data for that same as:key. If new data is we use that as main and only complement with old left over data. Which one if we deleted that just now? Well, we can actually have URLS and EXTERNAL files not managed by us. But. That also needs to be mimic in the WEBFORM because that little rascal gets rid of those.

ALSO: only reason we have that processing on the webform is because of ESIE. We may totally want to disable that option and let this part deal with everything

e.g