Add a general 720 seconds timer to stop the process
and make each queue take only 60 seconds max. We call this way many times the queue but we leave time for the rest of the timers to interact inbetween.

Once all is done. We simple go done and set PID to 0. NO need to cancel that timer since it runs only once.