Fix wrong array_filter
And now i know why i do not like array filters.
The array filter was using the value, the value is not a machine name so all permissions where failing. This adds ARRAY_FILTER_USE_KEY we have the actual form mode.

Also. My Drupal (verified) DCS are putting long lines with method accessors in a second row. I hope this is the right thing.