Remove also from Key Value when Removed from Index/tracker
@giancarlobi this now also removes it from the Key value when unindexed. Hope  this helps with the issue of re-appearing/race-condition ghosts. But we can still do better if so. I personally feel we can also have a Manual clean up option in case this happens as final option.

A side note:

The way the Tracker/Indexing works is a bit erratic when reindexing. Lets say you mark all for reindex. and some SBF are missing, if for a set of (50) non exist, the batch will stop. If you keep reindexing it will continue. Just in case you wonder why for that case you end with a 80%. Pressing index after that will keep doing it and all the ones not found will be untracked already.