Implemetent StrawberryfieldEventSearchApiForeignRelationshipsMapping
@giancarlobi this fixes the issue with the failing EntityQuery by removing every entity reference property belonging to a SBF. That means no more Error messages.
Still, this does not solved the fact that once an ADO is removed or updated, any other ADO pointing to it is not updated.
For that i will make another different Event Subscriber that will use the Index itself to query for any Solr Field that has a SBF property path pointing to a dynamic entity reference and check if "nid" is present. That way we can make it faster than an entityQuery ever could. Only issue is we have to Document/enforce that people always index their "nid" too. But that is way easier.