A less strict Date parser in case of non compliant EDTF
This will now
try with EDTF
If that fails try to parse as ISO8601 (why? because transforming to Unix timestamp directly here will destroy a time zone offset)
 if that fails as a string (stuff like "Yesterday", "2022-01" etc will work!) and in both cases return a Unixtimestamp that will then be casted back to ISO8601
If nothing works we return FALSE and that gets filtered out

This came up today while having a webform format its date as complete ISO8601.