Fixing missing Snippets
The OCR returns were overwriting previous results!
@patdunlavey @giancarlobi this should now Match more things on a Book search.
// Still incomplete and needs more checking
(we need to be sure the new Solr Fields are always present)

One note:
If an OCR contains "Queen;" "Queen" won't match. I wonder if we need either add spacing on the OCR itself to allow this to match or we can alter the HOCR processor/fieldtype to tokenize against also ",", "-" and ";" ? (Solr config)