Make sure delete/save of a child also tracks the parent Nodes
This adds a bit of extra work but only way of ensuring changes to children's Flavors will permeate (and be aggregated) on a parent one. If aggregation is not enabled then nothing terrible will happen, we will still mark to be updated. Less computing that actually checking in ANY index and specially the INDEX for parents have an aggregated Field present (which would do much more harm performance wise)