Enabled backend Flavor Highlight IF sbf_join_flavor option is enabled
This also forces $hl->setRequireFieldMatch(TRUE); to avoid highlighting things that did not match all. This is fair.

Most of this code depends on actual format_strawberry field views Filter (JOIN) one so needs to be tested together. There is more to do here but this looks clean. I could expose WAY more settings but for now we can roll with these defaults (e.g the snippets are going to be only of 128 max size

Finally i made a "if this is autocomplete" bypass highlighting.That should be by default Drupal, like really!