Ok. So... this is yet again a game changer
Previously we would fetch a SBF for a snippet, use its Item ID to get the parent ADO and use that ADO to match against the results so we know which result would get a snippet. Clearly when you have multiple hierarchies (and are hiding the actual directly connected objects from display) this fails. So now we use the returned values from the query/intersect them with the result list items and then generate the new snippets. Wonder how slower all this will be?